Move group and filter transforms to main
